IBX-6937: Changed maxFileSize value to float type for Image field type (#291)
For more details see https://issues.ibexa.co/browse/IBX-6937 and https://github.com/ibexa/core/pull/291

Changed max file size value from integer to float (numeric) type for Image field type. Storing the setting has been moved from dataInt1 to dataFloat1 relevant relation's column.

Key changes:

* Changed maxFileSize value type to float

* [DB] Updated schema

* Added max file size unit to storage dataText1 column

* Adjusted integration tests

* Changed field to store max file size from dataInt1 to dataFloat1

* Adjusted clean data to store max file size in dataFloat1 column

* Dropped converting max file size from MB to KB

---------

+++ tests/lib/FieldType/ImageTest.php
@@ -112,7 +112,7 @@ class ImageTest extends FieldTypeTest
         return [
             'FileSizeValidator' => [
                 'maxFileSize' => [
-                    'type' => 'int',
+                    'type' => 'numeric',
                     'default' => null,
                 ],
             ],
@@ -561,7 +561,7 @@ class ImageTest extends FieldTypeTest
                 [
                     'validatorConfiguration' => [
                         'FileSizeValidator' => [
-                            'maxFileSize' => 1,
+                            'maxFileSize' => 1.0,
                         ],
                     ],
                 ],
